Extract helper module section_label_helper.js from component VerseSelection
